+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 6.2.0
+* Add `chargebackProtectionLevel ` to `Dispute` and `DisputeSearch`
+* Add `skipAdvancedFraudChecking` to: 
+  * `CreditCard.create` and `CreditCard.update`
+  * `PaymentMethod.create` and `PaymentMethod.update`
+
 ## 6.1.0
 * Add `paypalMessages` to `Dispute`
 * Fix bug where `__isset` methods in `Instance` and `Base` classes treated `null` value as set (Thanks @sklodzio)